+++ src/senaite/patient/browser/patient/samples.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+
+from senaite.core.browser.samples.view import SamplesView as BaseView
+from senaite.patient import messageFactory as _
+
+
+class SamplesView(BaseView):
+    """Patient Samples View
+    """
+    def __init__(self, context, request):
+        super(SamplesView, self).__init__(context, request)
+
+        self.title = _("Samples of {}").format(
+            self.context.getFullname())
+
+        mrn = self.context.getMRN()
+        if mrn:
+            self.contentFilter["medical_record_number"] = [mrn]
+        else:
+            self.contentFilter["medical_record_number"] = []
+
+    def update(self):
+        """Called before the listing renders
+        """
+        super(SamplesView, self).update()